Added model validation + created validators.py
