+++ client/forms.py
@@ -1,16 +1,13 @@
 import json
-import re
 
 from django import forms
 from django.core.validators import RegexValidator
-from django.core.exceptions import ValidationError
 
 from client.models import Language, DEFAULT_MAX_FIELD_LENGTH, ID_LENGTH
+from client.validators import id_number_validator
 
 FIELD_NAME_MAPPING = {
 }
-
-ID_NUMBER_REGEX = r'^\d{8,9}$'
 
 AREAS = (
     ("צפון", "צפון"),
@@ -35,29 +32,6 @@ json_file.close()
 
 def get_lang_choices():
     return [(str(x), str(x)) for x in Language.objects.all()]
-
-
-def id_number_validator(value):
-    def digits_of(n):
-        return [int(d) for d in str(n)]
-
-    if not re.match(ID_NUMBER_REGEX, value):
-        raise ValidationError('Invalid ID number.', params={'value': value})
-    if len(value) == 8:
-        value = '0' + value
-
-    # Calculate Sifrat Bikoret (Luhn Checksum)
-    checksum = 0
-    digits = digits_of(value)
-    odd_digits = digits[:-1:2]
-    even_digits = digits[1:-1:2]
-    checksum += sum(odd_digits)
-    for digit in even_digits:
-        checksum += sum(digits_of(digit * 2))
-    checksum = 10 - (checksum % 10)  # Get difference from upper round number
-
-    if str(checksum) != value[-1]:
-        raise ValidationError('Invalid ID number: Incorrect check digit')
 
 
 class VolunteerForm(forms.Form):
